backend/requirements_lite.txt
# Requirements optimisés - compatibles Render plan gratuit
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0

# Géospatial léger (sans GDAL/fiona)
pyshp==2.3.1
shapely==2.0.4

# Base de données
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Données
pandas==2.2.2
numpy==1.26.4
geojson==3.1.0

# Export cartes
matplotlib==3.8.4
reportlab==4.2.0
pillow==10.3.0

# Auth & utilitaires
pyjwt==2.8.0
requests==2.32.2
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.7.1
email-validator==2.1.1
